+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bug fixes:
 
 ### Dependencies:
+- build(deps): `golang.org/x/net` from 0.43.0 to 0.44.0 ([#1538](https://github.com/fastly/cli/pull/1538))
 
 ## [v12.0.0](https://github.com/fastly/cli/releases/tag/v12.0.0) (2025-09-10)